Run update automatically when plugin is first installed
"It sounds a bit mad it's not that dirty in practice." -- @Seldaek

Listen for package install events and if the installed package is our
plugin set a flag for later. When the post-install/update event fires
and the flag is set, create and run a second Composer\Installer
instance in update mode. This will allow the newly installed plugin an
opportunity to process directives in the current composer.json file.

Fixes #34