Better support for initial install of plugin
Capture additional state information about autoloader dump settings and
read information from the root config so that when our plugin is
initially installed we can do a more complete job of executing a second
install cycle. Ideally we would be able to copy this and other
information from the currently executing \Composer\Installer instance
but unfortunately the plugin does not have access to it via the event
system.

Closes #41