Add support for 'last defined package wins' conflict resolution
Introduce a new `extra.merge-plugin.replace` configuration directive
that can be used to change the default conflict resolution mechanism.
When enabled, duplicate package declarations in merged files will
overwrite the declarations made in earlier files. Files are loaded in
the order specified in `extra.merge-plugin.include` with globbed files
being loaded in alphabetical order.

This behavior can be used to replace a default stable package with a dev
or beta version of the same package for testing.

Closes #45