Add branch alias for master branch
Having a branch alias allows us to associate a semantic version number
with the development branch. This can be used by a consuming project to
follow the latest unreleased version while still retaining the ability
to prevent adoption of major/minor version breaks.

This version number should be bumped concurrently with or just before
pulling new minor version changing commits into the master branch. The
current master branch is progressing towards the 1.3.0 release milestone
so we will start with "1.3.x" as the alias. This can remain until we
start to bring in changes for the 1.4.0 release.

More details at <https://getcomposer.org/doc/articles/aliases.md>