Handle RootAliasPackage as correctly as we can
When the top level Package includes an "extra.branch-alias" directive
Composer wraps the normal `RootPackage` in a `RootAliasPackage`. This is
done to allow the `RootAliasPackage` to modify the version information
in "require", "require-dev", "conflict", "provide" and "replace" sections
of the package. Unfortunately `RootAliasPackage` does not implement the
complete public API of `RootPackage` so we need to apply special checks
for its presence.

Previously we tried to handle this by immediately unwrapping the aliased
root package and making all operations use it directly. This
unfortunately doesn't work as internally Composer caches data for the
collections that can be modified by the alias in the `RootAliasPackage`
at construction time. This prevents our changes to those collections
from being acted on at runtime.

We fix propagation of "require" and "require-dev" configuration by
operating on the `RootAliasPackage` directly again. For other collection
changing operations we continue to modify the unwrapped package. These
changes should propagate for all but the "conflict", "provide" and
"replace" collections which are cached by the `RootAliasPackage` and not
exposed with setters that allow post-construction modification. If an
aliased package tries to merge these sections we emit a warning via
Composer and continue gracefully.

Closes #62
Co-authored-by: Alexander Polyanskikh <alexander.polyanskikh@claromentis.com>