Update RootPackageInterface and RootAliasPackage handling
Composer added enhanced RootPackageInterface and RootAliasPackage
definitions in composer/composer@599ad77 that allow us to work directly
with RootPackageInterface instead of unwrapping the inner RootPackage.
The prior ExtraPackage::castRootPackage() method is renamed to
ExtraPackage::unwrapIfNeeded() and modified to only unwrap if the needed
method is missing from the root package. This backwards compatibility
will need to be retained until a stable Composer version is released.

Closes #83