Fix unit tests for Composer upstream changes
Composer was changed in https://github.com/composer/composer/pull/4542
to provide a full interface for RootAliasPackage so that we no longer
need to try and unwrap the aliased package. Update the unit tests to
detect the Composer version in use by looking for
a RootAliasPackage::setRepositories() method and adjust mock
expectations accordingly.