Always prepend repositories
Follow up to 9645a93 that makes the use of
RepositoryManager::prependRepository() the default behavior for all
repository merges. This ensures that repos added by the
included/required config files to have a higher priority than the repos
added in the root composer.json and most importantly higher than the
default packagist repo.

Closes #95